changed the biphase bw command to biphase_clk_speed so that setting the mpsse clock speed and setting the biphase bw are two different things. Removed any dependency of the code on synclink

+++ common/sip.c
@@ -628,7 +628,7 @@ void WatchPort(void* parameter)
     unsigned char extdat[256];
 
     char tname[6];
-    snprintf(tname, sizeof(tname), "COMM%1d", port + 1);
+    snprintf(tname, sizeof(tname), "COMM%1d", (int) (port + 1));
     nameThread(tname);
     // blast_startup("WatchPort startup\n");
